# Optimize zipper majority vote