--- a/pkg/types/types.go
+++ b/pkg/types/types.go
@@ -318,36 +318,38 @@ func areFloatsExactlyEqual(a, b float64) bool {
 	return a == b
 }
 
-func getPointMajorityValue(values []float64, equalityFunc floatEqualityFunc) (majorityValue float64, majorityCount int, err error) {
+func getPointMajorityValue(values []float64, equalityFunc floatEqualityFunc) (majorityValue float64, isMajority bool, err error) {
 	valuesCount := len(values)
 	if valuesCount == 0 {
-		return 0, 0, errors.New("no value for majority voting")
+		return 0, false, errors.New("no value for majority voting")
 	}
 
-	sort.Slice(values, func(i, j int) bool {
-		return values[i] < values[j]
-	})
-
-	majorityCount = 1
-	majorityValue = values[0]
-
-	currentCount := 1
-	currentValue := values[0]
-	for i := 1; i < len(values); i++ {
-		v := values[i]
-		if equalityFunc(v, currentValue) {
-			currentCount++
-			if currentCount > majorityCount {
-				majorityCount = currentCount
-				majorityValue = v
-			}
+	// Boyer–Moore majority vote algorithm
+	i := 0
+	var m float64
+	for _, v := range values {
+		if i == 0 {
+			m = v
+		}
+		if equalityFunc(m, v) {
+			i++
 		} else {
-			currentValue = v
-			currentCount = 1
+			i--
+		}
+		if i > valuesCount/2 {
+			return m, isMajority, nil
 		}
 	}
 
-	return majorityValue, majorityCount, nil
+	// check if m is the majority value
+	var majorityCount int
+	for _, v := range values {
+		if equalityFunc(v, m) {
+			majorityCount++
+		}
+	}
+
+	return m, majorityCount > valuesCount/2, nil
 }
 
 func mergeMetrics(metrics []Metric, replicaMismatchConfig cfg.RenderReplicaMismatchConfig) (metric Metric, stats MetricRenderStats) {
@@ -405,7 +407,7 @@ func mergeMetrics(metrics []Metric, replicaMismatchConfig cfg.RenderReplicaMisma
 				pointExists = true
 			}
 
-			if !equalityFunc(metric.Values[i], m.Values[i]) {
+			if !mismatchObserved && !equalityFunc(metric.Values[i], m.Values[i]) {
 				mismatchObserved = true
 				if replicaMatchMode == cfg.ReplicaMatchModeCheck {
 					// mismatch exists, enough for check mode
@@ -419,8 +421,8 @@ func mergeMetrics(metrics []Metric, replicaMismatchConfig cfg.RenderReplicaMisma
 
 		mismatches++
 		if replicaMatchMode == cfg.ReplicaMatchModeMajority {
-			majorityValue, majorityCount, err := getPointMajorityValue(valuesForPoint, equalityFunc)
-			if err == nil && majorityCount > len(valuesForPoint)/2 {
+			majorityValue, isMajority, err := getPointMajorityValue(valuesForPoint, equalityFunc)
+			if err == nil && isMajority {
 				metric.Values[i] = majorityValue
 				fixedMismatches++
 			}