@bom-d-van 2022-03-15
tips: https://pkg.go.dev/golang.org/x/tools/cmd/benchcmp could be handy here, it prints out delta changes that's more readable.

```
benchcmp old.txt new.txt
benchcmp is deprecated in favor of benchstat: https://pkg.go.dev/golang.org/x/perf/cmd/benchstat
benchmark                                                                                         old ns/op       new ns/op       delta
BenchmarkRendersStorm/BenchmarkRendersStorm/ReplicaMatchMode-normalWithExactEqCheck-8             6052688709      6213874116      +2.66%
BenchmarkRendersStorm/BenchmarkRendersStorm/ReplicaMatchMode-checkWithExactEqCheck-8              7349380146      7012712732      -4.58%
BenchmarkRendersStorm/BenchmarkRendersStorm/ReplicaMatchMode-checkWithApproximateEqCheck-8        10030765581     10050420245     +0.20%
BenchmarkRendersStorm/BenchmarkRendersStorm/ReplicaMatchMode-majorityWithExactEqCheck-8           33215526638     20160802573     -39.30%
BenchmarkRendersStorm/BenchmarkRendersStorm/ReplicaMatchMode-majorityWithApproximateEqCheck-8     39880077185     31473745254     -21.08%
```